rename `define/ui` to `define/dangerously-all-ui`
